ingest/requirements.txt
psycopg[pool,c]==3.1.19
pyzmq~=26.0.3
redis[hiredis]~=5.0.4
